requirements.txt
nibabel==3.1.1
numpy
pyradiomics==3.0.1
scikit_image==0.15.0
nilearn==0.7.0
matplotlib==3.3.4
pandas==1.1.4
MedPy==0.4.0
Keras==2.2.5
h5py==2.10.0
git+https://www.github.com/keras-team/keras-contrib.git